Move parser tests to a separate file
